+++ test/functional/friends.sql
